Add portfolio validation CLI
